# src/main/resources/static/swagger.yaml
openapi: 3.0.0
info:
  version: 0.0.1
  title: Health Information User
  description: >
    "Health information User" refers to an entity that wishes to consume the
    services of the Health Data Consent Manager and obtain aggregated health
    information for providing services to the patient .
servers:
  - url: https://ncg-dev.projecteka.in/hiu-api
    description: Dev
...
paths:
...
  /users:
    post:
      tags:
        - frontend
      summary: Create user
      description: Create HIU user who can access HIU application
      parameters:
        - $ref: '#/components/parameters/authorization'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserCreationRequest'
      responses:
        '200':
          description: Successful creation of consent artefact.
        '401':
          description: Unauthorized request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Does not have enough permission to create user.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /users/password:
    put:
      tags:
        - frontend
      summary: Change password
      description: Change the password for logged in user
      parameters:
        - $ref: '#/components/parameters/authorization'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangePasswordBody'
      responses:
        '200':
          description: Successful updation of password.
        '400':
          description: When new password is invalid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  parameters:
    authorization:
      name: Authorization
      in: header
      required: true
      description: Bearer token which was issued after successful login.
      schema:
        type: string
...
  schemas:
...
    ChangePasswordBody:
      type: object
      properties:
        oldPassword:
          type: string
          example: someOldpassword@1
        newPassword:
          type: string
          example: someNewpassword@1
...
    ErrorResponse:
      type: object
      properties:
        error:
          $ref: '#/components/schemas/Error'
      xml:
        name: ErrorResponse
    Error:
      type: object
      properties:
        code:
          type: integer
          enum: [1000, 1001, 1007]
          description: >
            1. Error code 1000 : No patient found
            2. Error code 1001: Multiple patients found
        message:
          type: string
      xml:
        name: Error
...
    UserCreationRequest:
      type: object
      required:
        - username
        - password
      properties:
        username:
          type: string
        password:
          type: string
        role:
          type: string
          enum: [DOCTOR, ADMIN]
          default: DOCTOR
...